+++ demo/assets/extensions/datatables.net-bs5/types/dataTables.bootstrap5.d.ts
@@ -1,4 +1,5 @@
 
-import Api from "datatables.net";
+import DataTable from 'datatables.net';
 
-export default Api;
+export default DataTable;
+export * from 'datatables.net';